+++ app/components/AppAudioPlayer.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from 'react';
-import { Pressable, StyleSheet, ViewStyle } from 'react-native';
+import { Pressable, StyleSheet, ViewStyle, View } from 'react-native';
 import { MaterialIcons } from '@expo/vector-icons';
 import {
     Audio,
@@ -10,17 +10,18 @@ import {
 } from 'expo-av';
 
 import { COLORS } from '@config/colors';
-import { horizontalScale } from '@services/scale';
+import { horizontalScale, moderateScale, verticalScale } from '@services/scale';
+import PlaybackSpeedControl from './PlaybackSpeedControl';
 
 type Props = {
     sourceUri: string;
-    style: ViewStyle;
-    playbackSpeed: number;
 };
 
-function AppAudioPlayer({ sourceUri, style, playbackSpeed }: Props) {
+function AppAudioPlayer({ sourceUri }: Props) {
     const [isPlaying, setIsPlaying] = useState<boolean>(false);
     const { current: sound } = useRef(new Audio.Sound());
+    const [playbackSpeed, setPlaybackSpeed] = useState<number>(1.0);
+    const speeds = [1.0, 1.5, 2.0, 0.8];
 
     useEffect(() => {
         loadTrack();
@@ -76,14 +77,31 @@ function AppAudioPlayer({ sourceUri, style, playbackSpeed }: Props) {
         }
     }, [playbackSpeed]);
 
+    const handleSpeedChange = () => {
+        const currentIndex = speeds.indexOf(playbackSpeed);
+        const nextIndex = (currentIndex + 1) % speeds.length;
+        setPlaybackSpeed(speeds[nextIndex]);
+    };
+
     return (
-        <Pressable style={[styles.container, style]} onPress={handlePress}>
-            <MaterialIcons
-                name={isPlaying ? 'pause' : 'play-arrow'}
-                size={horizontalScale(30)}
-                color={COLORS.white}
-            ></MaterialIcons>
-        </Pressable>
+        <View style={[styles.controlsContainer]}>
+            <PlaybackSpeedControl
+                style={styles.playerContainer}
+                currentSpeed={playbackSpeed}
+                onPress={handleSpeedChange}
+            />
+
+            <Pressable
+                style={[styles.container, styles.playerContainer]}
+                onPress={handlePress}
+            >
+                <MaterialIcons
+                    name={isPlaying ? 'pause' : 'play-arrow'}
+                    size={horizontalScale(30)}
+                    color={COLORS.white}
+                ></MaterialIcons>
+            </Pressable>
+        </View>
     );
 }
 
@@ -94,6 +112,17 @@ const styles = StyleSheet.create({
         justifyContent: 'center',
         alignItems: 'center',
         backgroundColor: COLORS.primary,
+    },
+    controlsContainer: {
+        flexDirection: 'row-reverse',
+        justifyContent: 'flex-end',
+        marginBlockStart: verticalScale(16),
+        gap: 8,
+    },
+    playerContainer: {
+        width: horizontalScale(40),
+        aspectRatio: 1,
+        borderRadius: moderateScale(40),
     },
 });
